Change the hook for multithread. (#73)
Now multithread will always have a dummy profile function, 
so we can enable/disable the profiling anytime we want.

The downside is, as we will only remove the profile function
when viztracer is deallocated, it's possible that the timing of
garbage collection affects the behavior of viztracer. This won't
be a problem if there is only one tracer per process. However,
for tests it's not the case. We need to enforce a garbage
collection after each test.

This fix part of #72 when a spawned thread can't find thread
specific info without triggering profile function